Backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
bcrypt==4.2.1
    # via flask-bcrypt
certifi==2024.8.30
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via flask
cryptography==44.0.0
    # via -r requirements.in
flask==2.2.2
    # via
    #   -r requirements.in
    #   flask-bcrypt
    #   flask-mysql
flask-bcrypt==1.0.1
    # via -r requirements.in
flask-mysql==1.5.2
    # via -r requirements.in
idna==3.10
    # via requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via flask
kaleido==0.2.1
    # via -r requirements.in
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
mysql-connector-python==9.1.0
    # via -r requirements.in
numpy==2.1.3
    # via pandas
packaging==24.2
    # via plotly
pandas==2.2.3
    # via -r requirements.in
plotly==5.24.1
    # via -r requirements.in
pycparser==2.22
    # via cffi
pymysql==1.1.1
    # via flask-mysql
python-dateutil==2.9.0.post0
    # via pandas
pytz==2024.2
    # via pandas
requests==2.32.3
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
tenacity==9.0.0
    # via plotly
tzdata==2024.2
    # via pandas
urllib3==2.2.3
    # via requests
werkzeug==2.2.3
    # via
    #   -r requirements.in
    #   flask
